requirements.txt
keyboard
pyperclip
requests
langdetect
google-generativeai
pystray
pillow